+++ Tests/Functional/Model/RealtyObjectTest.php
@@ -5,7 +5,9 @@ namespace OliverKlee\Realty\Tests\Functional\Model;
 use Nimut\TestingFramework\TestCase\FunctionalTestCase;
 use Prophecy\Prophecy\ObjectProphecy;
 use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
+use TYPO3\CMS\Core\Resource\File;
 use TYPO3\CMS\Core\Resource\FileReference;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * Test case.
@@ -43,6 +45,31 @@ class RealtyObjectTest extends FunctionalTestCase
         $this->realtyObjectMapper = new \tx_realty_Mapper_RealtyObject();
 
         $this->subject = new \tx_realty_Model_RealtyObject();
+    }
+
+    protected function tearDown()
+    {
+        if (\file_exists($this->getAbsoluteAttachmentsTargetFolder())) {
+            GeneralUtility::rmdir($this->getAbsoluteAttachmentsTargetFolder(), true);
+        }
+
+        parent::tearDown();
+    }
+
+    /**
+     * @return string
+     */
+    private function getAbsoluteFixturesPath()
+    {
+        return __DIR__ . '/../Fixtures/';
+    }
+
+    /**
+     * @return string
+     */
+    public function getAbsoluteAttachmentsTargetFolder()
+    {
+        return GeneralUtility::getFileAbsFileName('fileadmin/realty_attachments');
     }
 
     /**
@@ -131,9 +158,9 @@ class RealtyObjectTest extends FunctionalTestCase
     {
         $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
 
-        /** @var \tx_realty_Model_RealtyObject $model */
-        $model = $this->realtyObjectMapper->find(101);
-        $attachments = $model->getAttachments();
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(101);
+        $attachments = $subject->getAttachments();
 
         self::assertSame([], $attachments);
     }
@@ -146,9 +173,9 @@ class RealtyObjectTest extends FunctionalTestCase
         $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
         $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
 
-        /** @var \tx_realty_Model_RealtyObject $model */
-        $model = $this->realtyObjectMapper->find(102);
-        $attachments = $model->getAttachments();
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+        $attachments = $subject->getAttachments();
 
         self::assertGreaterThanOrEqual(1, $attachments);
         $firstAttachment = $attachments[0];
@@ -164,9 +191,9 @@ class RealtyObjectTest extends FunctionalTestCase
         $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
         $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
 
-        /** @var \tx_realty_Model_RealtyObject $model */
-        $model = $this->realtyObjectMapper->find(102);
-        $attachments = $model->getPdfAttachments();
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+        $attachments = $subject->getPdfAttachments();
 
         self::assertCount(1, $attachments);
         $firstAttachment = $attachments[0];
@@ -182,13 +209,200 @@ class RealtyObjectTest extends FunctionalTestCase
         $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
         $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
 
-        /** @var \tx_realty_Model_RealtyObject $model */
-        $model = $this->realtyObjectMapper->find(102);
-        $attachments = $model->getJpegAttachments();
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+        $attachments = $subject->getJpegAttachments();
 
         self::assertCount(1, $attachments);
         $firstAttachment = $attachments[0];
         self::assertInstanceOf(FileReference::class, $firstAttachment);
         self::assertSame('test.jpg', $firstAttachment->getName());
+    }
+
+    /**
+     * @test
+     *
+     * @expectedException \BadMethodCallException
+     */
+    public function addAndSaveAttachmentForVirginModelThrowsException()
+    {
+        $subject = new \tx_realty_Model_RealtyObject();
+
+        $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . 'test2.jpg', 'test image');
+    }
+
+    /**
+     * @test
+     *
+     * @expectedException \BadMethodCallException
+     */
+    public function addAndSaveAttachmentForEmptyModelWithoutUidThrowsException()
+    {
+        $subject = new \tx_realty_Model_RealtyObject();
+        $subject->setData([]);
+
+        $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . 'test2.jpg', 'test image');
+    }
+
+    /**
+     * @test
+     *
+     * @expectedException \UnexpectedValueException
+     */
+    public function addAndSaveAttachmentForInexistentFileThrowsException()
+    {
+        $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
+
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+        $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . 'there-is-no-image.jpg', 'test image');
+    }
+
+    /**
+     * @test
+     */
+    public function addAndSaveAttachmentForCompletelyNewAttachmentsIncreasesNumberOfAttachments()
+    {
+        $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
+
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+        $oldNumberOfAttachments = $subject->getNumberOfAttachments();
+
+        $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . 'test2.jpg', 'test image');
+
+        self::assertSame($oldNumberOfAttachments + 1, $subject->getNumberOfAttachments());
+    }
+
+    /**
+     * @test
+     */
+    public function addAndSaveAttachmentReturnsForNewAttachmentReturnsPersistedFile()
+    {
+        $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
+
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+
+        $result = $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . 'test2.jpg', 'test image');
+
+        self::assertInstanceOf(File::class, $result);
+        self::assertGreaterThan(0, $result->getUid());
+    }
+
+    /**
+     * @test
+     */
+    public function addAndSaveAttachmentReturnsForNewAttachmentCopiesFileToObjectSpecificLocation()
+    {
+        $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
+
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+
+        $result = $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . 'test2.jpg', 'test image');
+
+        self::assertSame('/realty_attachments/102/test2.jpg', $result->getIdentifier());
+    }
+
+    /**
+     * @test
+     */
+    public function addAndSaveAttachmentReturnsForNewAttachmentSetsTitle()
+    {
+        $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
+
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+
+        $result = $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . 'test2.jpg', 'test image');
+
+        $metaData = \Tx_Oelib_Db::selectSingle('*', 'sys_file_metadata', 'file = ' . $result->getUid());
+        self::assertSame('test image', $metaData['title']);
+    }
+
+    /**
+     * @test
+     */
+    public function addAndSaveAttachmentCreatesReferenceToNewFile()
+    {
+        $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
+
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+
+        $result = $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . 'test2.jpg', 'test image');
+
+        $exists = false;
+        foreach ($subject->getAttachments() as $reference) {
+            $exists = $exists || ($reference->getOriginalFile()->getUid() === $result->getUid());
+        }
+        self::assertTrue($exists);
+    }
+
+    /**
+     * @test
+     */
+    public function addAndSaveAttachmentForDuplicatedFileNameFileReusesExistingFile()
+    {
+        $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
+        $targetFolder = $this->getAbsoluteAttachmentsTargetFolder() . '102/';
+        GeneralUtility::mkdir_deep($targetFolder);
+        \copy($this->getAbsoluteFixturesPath() . 'test2.txt', $targetFolder . 'test2.txt');
+
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+
+        $result = $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . '/test2.txt', 'test image');
+
+        self::assertSame('test2.txt', $result->getName());
+        self::assertSame('/realty_attachments/102/test2.txt', $result->getIdentifier());
+        self::assertSame(13, $result->getUid());
+    }
+
+    /**
+     * @test
+     */
+    public function addAndSaveAttachmentForDuplicatedFileNameFileUpdatesTitle()
+    {
+        $newTitle = 'new title';
+
+        $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
+        $targetFolder = $this->getAbsoluteAttachmentsTargetFolder() . '102/';
+        GeneralUtility::mkdir_deep($targetFolder);
+        \copy($this->getAbsoluteFixturesPath() . 'test.jpg', $targetFolder . 'test.jpg');
+
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+
+        $result = $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . '/test.jpg', $newTitle);
+
+        $metaData = \Tx_Oelib_Db::selectSingle('*', 'sys_file_metadata', 'file = ' . $result->getUid());
+        self::assertSame($newTitle, $metaData['title']);
+    }
+
+    /**
+     * @test
+     */
+    public function addAndSaveAttachmentForDuplicatedFileNameFileReusesExistingFileReference()
+    {
+        $this->importDataSet(__DIR__ . '/../Fixtures/Attachments.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/RealtyObjects.xml');
+
+        /** @var \tx_realty_Model_RealtyObject $subject */
+        $subject = $this->realtyObjectMapper->find(102);
+
+        $subject->addAndSaveAttachment($this->getAbsoluteFixturesPath() . 'test.jpg', 'test image');
+
+        $numberOfReferences = \Tx_Oelib_Db::count('sys_file_reference', 'uid_local = 10 AND uid_foreign = 102');
+        self::assertSame(1, $numberOfReferences);
     }
 }